% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDisp.par.R
\name{estimateDisp.par}
\alias{estimateDisp.par}
\title{Parallelized version of \code{edgeR::estimateDisp} function.}
\usage{
estimateDisp.par(
  y,
  design = NULL,
  group = NULL,
  lib.size = NULL,
  offset = NULL,
  prior.df = NULL,
  trend.method = "locfit",
  tagwise = TRUE,
  span = NULL,
  min.row.sum = 5,
  grid.length = 21,
  grid.range = c(-10, 10),
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  tol = 1e-06,
  weights = NULL,
  BPPARAM = BPPARAM,
  ...
)
}
\description{
Parallelized version of \code{edgeR::estimateDisp} function.
}
