% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSCE.R
\name{makeSCE}
\alias{makeSCE}
\title{Create SingleCellExperiment object containing normalized data}
\usage{
makeSCE(obj, cData = NULL, batch = NULL)
}
\arguments{
\item{obj}{object containing output of call to ruvIII.nb or fastruvIII.nb function.}

\item{cData}{A data frame containing cell-level metadata. This data frame will be used as 'colData' in the SingleCellExperiment object}

\item{batch}{numeric vector containing batch information for each sample. Must correspond to columns of count matrix. Only needed if batch-specific dispersion parameter is fitted.}
}
\value{
A SingleCellExperiment object with normalized data added to the 'assays' slot. The log percentile-adjusted count is stored in the 'logcorrected' component of the 'assays' slot and the Pearson residuals is in the 'pearson' component.
}
\description{
This function takes ruvIII.nb or fastruvIII.nb output as input and creates a SingleCellExperiment (SCE) object containing various metrics of normalized data.
}
