% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvIIInb.R
\name{ruvIII.nb}
\alias{ruvIII.nb}
\title{Removing unwanted variation from sequencing count data}
\usage{
ruvIII.nb(
  Y,
  M,
  ctl,
  k = 2,
  robust = FALSE,
  ortho.W = FALSE,
  lambda.a = 0.01,
  lambda.b = 16,
  batch = NULL,
  step.fac = 0.5,
  inner.maxit = 50,
  outer.maxit = 25,
  ncores = 2,
  use.pseudosample = FALSE,
  nc.pool = 20,
  batch.disp = FALSE,
  zeroinf = NULL
)
}
\arguments{
\item{Y}{raw count matrix with genes as rows and cells as columns}

\item{M}{replicate matrix with number of rows equal to the number of cells and number of columns equal to the number of distinct sub-population of cells: M(i,j) = 1 if cell i is part of replicate j and 0 otherwise. If cell-type is known, this can be used to form the replicate matrix, with each cell-type corresponds to one sub-population of cells. If cell-type is unknown, the scMerge::scReplicate command can be used to identify the pseudo-replicates.}

\item{ctl}{either logical vector of length n with TRUE = control genes OR a character vector of names of control genes}

\item{k}{dimension of unwanted variation against which the data is to be normalized against. Default is 2.}

\item{robust}{logical value. Whether to use Huber's robust weight in the iterative reweighted least squares (IRLS) algorithm. Default is FALSE.}

\item{ortho.W}{logical value. Whether the unwanted factors (W) are to be orthogonalized. The default is FALSE.}

\item{lambda.a}{smoothing parameter for regularizing regression coefficients associated with W. The default is 0.01}

\item{lambda.b}{smoothing parameter for regularizing the gene-level intercepts. The default is 16.}

\item{batch}{numeric vector containing batch information for each cell. Must correspond to columns of count matrix. If not supplied, cells are assumed to come from one batch.}

\item{step.fac}{multiplicative factor to decrease IRLS step by when log-likelihood diverges. Default is 0.5}

\item{inner.maxit}{the maximum number of IRLS iteration for estimating mean parameters for a given dispersion parameter. Default is 50}

\item{outer.maxit}{the maximum number of IRLS iteration. Default is 25}

\item{ncores}{The number of cores used for parallel processing. The default number of workers (cores) is 2.}

\item{use.pseudosample}{whether to use pseudocell to define replicates (default is FALSE). Note that the replicates defined by the pseudocells will be used in addition to any replicates defined by the M matrix above. We recommend that use.pseudosample=TRUE be used only for data with UMI.}

\item{nc.pool}{the number of cells per pool (used for defining pseudocells). The default is 20. Only relevant when use.pseudosample=TRUE.}

\item{batch.disp}{whether to fit batch-specific dispersion parameter for each gene. The default is FALSE.}

\item{zeroinf}{logical vector indicating whether to fit zero-inflated negative binomial (ZINB) instead of NB model (the default) for some/all cells. The length of the vector #'        must equal the number of cells. We recommend that non-UMI data be fitted using ZINB and UMI data be fitted using NB model.}
}
\value{
A list containing the raw data (as sparse matrix), the unwanted factors, regression coefficients associated with the unwanted factors, the M matrix and the estimates of dispersion parameters
}
\description{
This function performs ruvIIInb method to remove unwanted variation from sequencing count data. The method has been applied to single-cell RNA-seq data but in principle it is applicable to count data from any sequencing platforms. It takes raw count matrix with features (genes/transcript in scRNA-seq) as rows and samples (cells in scRNA-seq) as columns.
Users need to specify the set of negative control genes (i.e genes where the between cells variation is assumed to be solely due to the unwanted variation)
and the (pseudo)replicate matrix that define sets of cells that can be considered technical replicates.
}
