% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.res.R
\name{get.res}
\alias{get.res}
\title{Produce normalized count data after adjusting for unwanted variations}
\usage{
get.res(out, type = c("logcounts", "pearson", "quantile"), batch = NULL)
}
\arguments{
\item{out}{output of call to ruvIII.nb or fastruvIII.nb function.}

\item{type}{type of normalized data. Supported options are 'quantile' (percentile-adjusted count), 'pearson' (pearson residuals) and logcounts (log normalized count).}

\item{batch}{numeric vector containing batch information for each sample.Must correspond to columns of count matrix. Only needed if batch-specific dispersion parameter is fitted.}
}
\value{
A matrix containing the normalized data
}
\description{
This internal function takes ruvIII.nb or fastruvIII.nb output as input and produce various metrics of normalized data.  Option for metrics of normalized data includes percentile-adjusted count (PAC) and Pearson residuals.
}
