% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSCE.R
\name{makeSCE}
\alias{makeSCE}
\title{Add metrics of normalized data to SingleCellExperiment object}
\usage{
makeSCE(obj, cData, batch = NULL)
}
\arguments{
\item{obj}{An ruvIII.nb output.}

\item{cData}{A data frame containing cell-level metadata.}

\item{batch}{numeric vector containing batch information for each cell. Must corresponds to columns of count matrix. Only needed if batch-specific dispersion parameter is fitted.}
}
\value{
A SingleCellExperiment object with metrics of normalized data in the assay slot. The \code{pearson} component of the assay slot contains the pearson residuals and the \code{logcorrected} component contains the log of percentile-invariant adjusted data + 1.
}
\description{
This function takes ruvIIInb output and creates SingleCellExperiment object that contains various metrics of normalized data.
}
